Fix warnings for gcc and partialy for VS

+++ apps/freeablo/main.cpp
@@ -375,7 +375,7 @@ void runGameLoop(const bpo::variables_map& variables)
 
     FAWorld::World world;
 
-    FALevelGen::FAsrand(time(NULL));
+    FALevelGen::FAsrand((int)time(NULL));
 
     std::vector<Level::Level*> levels(13);
